+++ packages/contracts/contracts/interfaces/IDepositChallenge.sol
@@ -1,5 +1,5 @@
 pragma solidity >= 0.6.0;
 
 interface IDepositChallenge {
-    function challengeMassDeposit(uint index, bytes calldata) external;
+    function challengeMassDeposit(uint256 index, bytes calldata) external;
 }
